requirements.txt
data==0.4
decorator==5.1.1
funcsigs==1.0.2
future==0.18.3
lark==1.1.9
latex==0.7.0
PyYAML==6.0.1
shutilwhich==1.1.0
six==1.16.0
tempdir==0.7.1
